(#39) Home 컴포넌트 코드 최적화
